+++ py/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add protocol support for custom units on transaction measurements. ([#1256](https://github.com/getsentry/relay/pull/1256))
+- Add a profile data category and count profiles in an envelope to apply rate limits. ([#1259](https://github.com/getsentry/relay/pull/1259))
 
 ## 0.8.10